No BetaReductionVisitor
Using an observer to actually change the model is not a good idea,
because observer are notified in a random order.

Essentially, beta reduction needs to find a tree pattern like the one
below and modify it.

        X                      X
        |                      |
       App        =beta=>      Z'
      /   \
    Abs    Y
     |
     Z

For pattern matching, we are not allowed to use `instanceof`.
We could use `isApplication`, but that is just the same.
Proper object-oriented means just sending messages.
So, we just send `oneBetaReduction` message to children.
Applications receiving such a message,
message `apply(left)` to their right hand side.
Abstractions receiving `apply` perform the application.

Implementing beta reduction via visitor does not fit this very well.
For example, we need return values to find out,
if a reduction happened and if the current object (consider X above)
needs to exchange some its children.
We do not want reference to parent objects.